Refactor: model based and meta rl
