+++ setup.py
@@ -4,7 +4,8 @@ from setuptools import setup
 setup(
     name='unstable_baselines',
     auther='lamda-5',
-    packages=find_packages(),
+    # packages=find_packages(),
+    packages=[package for package in find_packages() if package.startswith("unstable_baselines")], 
     python_requires='>=3.7',
     package_data={
         # include default config files and env data files
